requirements.txt
blend_modes==2.1.0
numpy==1.24.1
openai==0.26.5
opencv_python==4.7.0.68
Pillow==9.4.0
PyAutoGUI==0.9.53
selenium==4.8.2
